🎉 Woot woot! Looks like we got a fancy little change to the SSH config! 🎊
Here's the deets:
- Added a new host config for `192.168.200.8` 🖥️
  - Specified the `HostName` as `192.168.200.8` (in case that wasn't obvious 😅)
  - Set the `User` to `harper` (VIP access! 😎)

And that's it! One shiny new host config, ready to rock and roll! 🚀

Let me break it down for any SSH n00bs out there:
- This lets you easily SSH into the server at `192.168.200.8` using the `harper` user account
- All your other default SSH settings still apply, so it's secure AF 🔒
- Now you can just run `ssh 192.168.200.8` and boom, you're in like Flynn! 😏

Hope this helps explain the change! If not, feel free to tell me to STFU and RTFM. 😘

Keep on keepin' on with that awesome SSH config! 🙌